Tests/LibWeb/Text/expected/Worker/Worker-offscreencanvas.txt

Message received from worker: "{\"w\":10,\"h\":10}"
Message received from worker: "20px SerenitySans"
Message received from worker: "20px SerenitySans"
